requirements.txt
click==8.0.1
configparser==5.0.2
Flask==2.0.1
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.8.6
python-dotenv==0.17.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1